requirements.txt
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1734469393482/work
aiohttp @ file:///croot/aiohttp_1734687138658/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///croot/async-timeout_1732661977001/work
attrs @ file:///croot/attrs_1734533101012/work
beautifulsoup4==4.13.3
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
craft-text-detector==0.4.3
datasets @ file:///croot/datasets_1716911606380/work
dill @ file:///croot/dill_1715094664823/work
filelock @ file:///croot/filelock_1700591183607/work
frozenlist @ file:///croot/frozenlist_1730902802621/work
fsspec @ file:///croot/fsspec_1714461537038/work
gdown==5.2.0
huggingface_hub @ file:///croot/huggingface_hub_1724853938404/work
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath==1.3.0
multidict @ file:///croot/multidict_1730905498140/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
networkx==3.2.1
numexpr @ file:///croot/numexpr_1730215937391/work
numpy==1.23.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.5.4.60
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp39-cp39-linux_x86_64.whl#sha256=0a51ed2e81ab863e3d00ed6c5049192ce578ecb38fb467d2f9a6585d3c25f666
pillow @ file:///croot/pillow_1738010226202/work
propcache @ file:///croot/propcache_1732303986938/work
pyarrow @ file:///croot/pyarrow_1738229418212/work/python
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
pytesseract @ file:///croot/pytesseract_1736867616419/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1730999120400/work
safetensors @ file:///croot/safetensors_1732227615904/work
scipy==1.13.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.6
sympy==1.13.1
tokenizers @ file:///croot/tokenizers_1732129617059/work
torch==1.12.0
torchvision==0.13.0
tqdm @ file:///croot/tqdm_1738943501192/work
transformers @ file:///croot/transformers_1729767158381/work
triton==3.2.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1737133630106/work
xxhash @ file:///croot/python-xxhash_1737039434400/work
yarl @ file:///croot/yarl_1732546845924/work
